Use unique key per-tilemap to avoid conflicts.
Previously, loading multiple tilemaps with the same external tileset
failed because the loader didn't bother adding the tileset more than
once since it already loaded a file with the same key; this left the
tilemap file hanging as it waited for the tileset to load.

Making tileset keys include the tilemap name in their key works around
this, and as long as your server is sending correct cache headers, the
browser should take care of not re-downloading the same tileset.